refactor(buffer): extract search types to search.rs
Moves ChunkedSearchState, HybridSearchPlan + impl (execute method),
SearchRegion enum, and the search_boundary_overlap free fn from
mod.rs into model/buffer/search.rs. These are standalone types that
don't need impl TextBuffer — HybridSearchPlan::execute takes
&dyn FileSystem as a param, and ChunkedSearchState is passed around
by the scan methods.

The 12 impl TextBuffer search methods (find_next, search_scan_*,
search_hybrid_plan, etc.) remain in mod.rs as they need
self.get_text_range / self.get_text_range_mut.

mod.rs drops from 3,892 → 3,575 (−317 lines). search.rs is 344
lines. 127 buffer tests green.
